DB connection fix with getUserById API
